An algorithm has name
